kernel: x86_64: refactor multiboot module (#435)

+++ src/kernel/arch/x86_64/kmain.h
@@ -9,6 +9,6 @@
  *
  * @param addr the multiboot info address
  */
-void kmain(uint64_t addr) __asm__("kmain");
+void kmain(uintptr_t addr) __asm__("kmain");
 
 #endif
